修复一些问题

+++ server/app/api/request/common.go
@@ -8,5 +8,5 @@ type PageInfo struct {
 
 // Find by id structure
 type GetById struct {
-	Id float64 `p:"id" v:"required|length:1,1000#请输入id|id长度为:min到:max位"`
+	Id int `p:"id" v:"required|length:1,1000#请输入id|id长度为:min到:max位"`
 }
